refactor(app-router): redirect to dashboard / home
